% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Displaying Prediction to Shiny Application}
\usage{
server(input, output)
}
\arguments{
\item{input}{Keep this as input, This will be using the input from the prediction model}

\item{output}{Keep this as output. This is what will be displayed to the user.}

\item{round}{Round the number that will be displayed}

\item{PredictionModel}{Calling the function where the user inputs are stored. Add here what your fuction is called.}

\item{digits}{Rounding numbers to the nearest digits after the decimal point.}
}
\description{
This function is used to display the prediction to the shiny application.
}
\examples{
(input, output)
(round((ForestModel(input)),digits=2))


}
