% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeModel.R
\name{TreeModel}
\alias{TreeModel}
\title{Adding UserInput into random forest}
\usage{
TreeModel(input)
}
\arguments{
\item{input}{This is referring to the input that will be given to the shiny application. Keep as input.}

\item{column_name<-input$coloum_name}{Set the column name that the user input will be going to.}

\item{inputList}{add the name of the varible you set the list of user inputs to be.}

\item{fit}{This is the name of tree fit that was created.}

\item{column~column}{Should include column that data is dependant on then columns to be predicted.}

\item{data=}{Name of the dataframe used should be after the =.}

\item{type=Vector}{Class or Vector}

\item{prediction}{This is the name of the variable you set your tree model prediction to be.}
}
\value{
The predicted house price based on user input
}
\description{
This function is adding user inputs into the random forest model
This is done by creating a list with all the user inputs making
them into a dataframe then feeding this dataframe into the model.
}
\examples{

fit <- rpart( column ~ column, data = dataframe_name, cp =.02)
inputList <- list(column_name<-input$Column_name,)
inputList <- as.data.frame(inputList) = predict(fit, inputList, type = "vector")
return(prediction)
}
